Refactor Swift extensions
