+++ pay_ios/ios/Classes/PaymentHandler.swift
@@ -208,172 +208,3 @@ extension PaymentHandler: PKPaymentAuthorizationControllerDelegate {
     }
   }
 }
-
-/// A set of utility methods associated to `PKPayment`.
-extension PKPayment {
-  
-  /// Creates a `Dictionary` representation of the `PKPayment` object.
-  public func toDictionary() -> [String: Any?] {
-    return [
-      "token": String(decoding: token.paymentData, as: UTF8.self),
-      "paymentMethod": token.paymentMethod.toDictionary(),
-      "billingContact": billingContact?.toDictionary(),
-      "shippingContact": shippingContact?.toDictionary(),
-      "shippingMethod": shippingMethod?.toDictionary(),
-    ]
-  }
-}
-
-/// A set of utility methods associated to `PKPaymentMethod`.
-extension PKPaymentMethod {
-  
-  /// Creates a `Dictionary` representation of the `PKPaymentMethod` object.
-  public func toDictionary() -> [String: Any?] {
-    return [
-      "displayName": displayName,
-      "network": network?.rawValue,
-      "type": type.rawValue
-    ]
-  }
-}
-
-/// A set of utility methods associated to `PKContact`.
-extension PKContact {
-  
-  /// Creates a `Dictionary` representation of the `PKContact` object.
-  public func toDictionary() -> [String: Any?] {
-    return [
-      "name": "\(name?.givenName ?? "") \(name?.familyName ?? "")",
-      "emailAddress": emailAddress,
-      "phoneNumber": phoneNumber?.stringValue
-    ]
-  }
-}
-
-/// A set of utility methods associated to `PKShippingMethod`.
-extension PKShippingMethod {
-  
-  /// Creates a `Dictionary` representation of the `PKShippingMethod` object.
-  public func toDictionary() -> [String: Any?] {
-    return [
-      "id": identifier,
-      "details": detail
-    ]
-  }
-}
-
-/// A set of utility methods associated to `PKPaymentNetwork`.
-extension PKPaymentNetwork {
-  
-  /// Creates a `PKPaymentNetwork` object from a network in string format.
-  public static func fromString(_ paymentNetworkString: String) -> PKPaymentNetwork? {
-    switch paymentNetworkString {
-    case "amex":
-      return .amex
-    case "cartesBancaires":
-      guard #available(iOS 12.0, *) else { return nil }
-      return .cartesBancaires
-    case "chinaUnionPay":
-      return .chinaUnionPay
-    case "discover":
-      return .discover
-    case "eftpos":
-      guard #available(iOS 12.0, *) else { return nil }
-      return .eftpos
-    case "electron":
-      guard #available(iOS 12.0, *) else { return nil }
-      return .electron
-    case "elo":
-      guard #available(iOS 12.1.1, *) else { return nil }
-      return .elo
-    case "idCredit":
-      return .idCredit
-    case "interac":
-      return .interac
-    case "JCB":
-      return .JCB
-    case "mada":
-      guard #available(iOS 12.1.1, *) else { return nil }
-      return .mada
-    case "maestro":
-      guard #available(iOS 12.0, *) else { return nil }
-      return .maestro
-    case "masterCard":
-      return .masterCard
-    case "privateLabel":
-      return .privateLabel
-    case "quicPay":
-      return .quicPay
-    case "suica":
-      return .suica
-    case "visa":
-      return .visa
-    case "vPay":
-      guard #available(iOS 12.0, *) else { return nil }
-      return .vPay
-    case "barcode":
-      guard #available(iOS 14.0, *) else { return nil }
-      return .barcode
-    case "girocard":
-      guard #available(iOS 14.0, *) else { return nil }
-      return .girocard
-    default:
-      return nil
-    }
-  }
-}
-
-/// A set of utility methods associated to `PKPaymentSummaryItemType`.
-extension PKPaymentSummaryItemType {
-  
-  /// Creates a `PKPaymentSummaryItemType` object from an item type in string format.
-  public static func fromString(_ summaryItemType: String) -> PKPaymentSummaryItemType {
-    switch summaryItemType {
-    case "pending":
-      return .pending
-    default:
-      return .final // final_price
-    }
-  }
-}
-
-/// A set of utility methods associated to `PKMerchantCapability`.
-extension PKMerchantCapability {
-  
-  /// Creates a `PKMerchantCapability` object from a capability in string format.
-  public static func fromString(_ merchantCapability: String) -> PKMerchantCapability? {
-    switch merchantCapability {
-    case "3DS":
-      return .capability3DS
-    case "EMV":
-      return .capabilityEMV
-    case "credit":
-      return .capabilityCredit
-    case "debit":
-      return .capabilityDebit
-    default:
-      return nil
-    }
-  }
-}
-/// A set of utility methods associated to `PKContactField`.
-extension PKContactField {
-  
-  /// Creates a `PKContactField` object from a contact field item in string format.
-  public static func fromString(_ contactFieldItem: String) -> PKContactField? {
-    switch contactFieldItem {
-    case "emailAddress":
-      return .emailAddress
-    case "name":
-      return .name
-    case "phoneNumber":
-      return .phoneNumber
-    case "phoneticName":
-      return .phoneticName
-    case "postalAddress":
-      return .postalAddress
-    default:
-      return nil
-    }
-  }
-}
